requirements.txt
Flask
requests
python-dotenv
gunicorn
openai
python-docx
azure-storage-blob
